Moved mock classes to new file
